@@ -10,6 +10,7 @@
     "url": "https://github.com/chia1104"
   },
   "scripts": {
+    "analyze": "cross-env ANALYZE=true npm run build",
     "dev": "next dev",
     "turbo": "concurrently \"next dev --turbo\" \"tailwindcss --input input.css --output output.css --watch\"",
     "build": "next build",
@@ -20,26 +21,29 @@
     "e2e:chrome": "cypress run --browser chrome",
     "e2e:firefox": "cypress run --browser firefox",
     "e2e:edge": "cypress run --browser edge",
-    "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\""
+    "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\"",
+    "type:check": "tsc --noEmit"
   },
   "dependencies": {
     "@geist-ui/core": "^2.3.8",
     "@giscus/react": "^2.2.6",
-    "@mdx-js/react": "2.2.1",
-    "@reduxjs/toolkit": "^1.9.2",
+    "@mdx-js/react": "2.3.0",
+    "@next/bundle-analyzer": "13.2.1",
+    "@radix-ui/react-accordion": "1.1.0",
+    "@reduxjs/toolkit": "^1.9.3",
     "@sendgrid/mail": "7.7.0",
-    "@vercel/analytics": "0.1.8",
-    "@vercel/og": "0.0.27",
-    "classnames": "^2.3.2",
+    "@vercel/analytics": "0.1.10",
+    "@vercel/og": "0.1.0",
+    "clsx": "1.2.1",
+    "cross-env": "7.0.3",
     "dayjs": "^1.11.7",
     "fast-glob": "^3.2.12",
-    "framer-motion": "^8.5.5",
+    "framer-motion": "^9.1.7",
     "graphql": "^16.6.0",
     "graphql-request": "^5.1.0",
     "gray-matter": "^4.0.3",
-    "next": "^13.1.6",
+    "next": "^13.2.1",
     "next-mdx-remote": "^4.3.0",
-    "next-seo": "^5.15.0",
     "next-themes": "^0.2.1",
     "p-map": "^5.5.0",
     "react": "^18.2.0",
@@ -48,48 +52,52 @@
     "react-redux": "^8.0.5",
     "reading-time": "^1.5.0",
     "rehype-autolink-headings": "^6.1.1",
-    "rehype-code-titles": "^1.1.0",
+    "rehype-code-titles": "^1.2.0",
     "rehype-highlight": "^6.0.0",
-    "rehype-prism-plus": "^1.5.0",
+    "rehype-prism-plus": "^1.5.1",
     "rehype-slug": "^5.1.0",
     "remark-gfm": "^3.0.1",
     "server-only": "0.0.1",
     "sharp": "0.31.3",
     "superjson": "1.12.2",
+    "tailwind-merge": "1.10.0",
     "usehooks-ts": "^2.9.1",
-    "zod": "^3.20.2"
+    "zod": "^3.20.6"
   },
   "devDependencies": {
     "@redux-devtools/extension": "^3.2.5",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/line-clamp": "^0.4.2",
-    "@testing-library/react": "13.4.0",
-    "@types/glob": "^8.0.1",
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.50.0",
-    "@typescript-eslint/parser": "^5.50.0",
-    "@vitejs/plugin-react": "^3.0.1",
+    "@testing-library/react": "14.0.0",
+    "@total-typescript/ts-reset": "0.3.7",
+    "@types/glob": "^8.1.0",
+    "@types/node": "^18.14.1",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@typescript-eslint/parser": "^5.53.0",
+    "@vitejs/plugin-react": "^3.1.0",
     "autoprefixer": "^10.4.13",
     "concurrently": "7.6.0",
-    "cypress": "^12.5.0",
-    "eslint": "^8.33.0",
-    "eslint-config-next": "^13.1.6",
+    "cypress": "12.6.0",
+    "eslint": "^8.34.0",
+    "eslint-config-next": "^13.2.1",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.3",
     "jsdom": "^21.1.0",
-    "lint-staged": "^13.1.0",
-    "next-sitemap": "^3.1.49",
+    "lint-staged": "^13.1.2",
+    "next-sitemap": "^3.1.52",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.3",
+    "prettier": "^2.8.4",
+    "prettier-plugin-tailwindcss": "0.2.3",
     "redux": "^4.2.1",
     "shell-quote": "^1.8.0",
     "tailwind-scrollbar": "^2.1.0",
-    "tailwindcss": "^3.2.4",
+    "tailwindcss": "^3.2.7",
+    "tailwindcss-animate": "1.0.5",
     "typescript": "4.9.5",
-    "vitest": "^0.28.3"
+    "vitest": "^0.28.5"
   },
   "engines": {
     "node": ">=16.x"
