@@ -1,7 +1,7 @@
 export default function PostLoading() {
   return (
-    <div className="flex justify-center items-center h-screen">
-      <div className="loader ease-linear rounded-full border-4 border-t-4 border-gray-200 h-12 w-12 mb-4" />
+    <div className="flex h-screen items-center justify-center">
+      <div className="loader mb-4 h-12 w-12 rounded-full border-4 border-t-4 border-gray-200 ease-linear" />
     </div>
   );
 }
