@@ -3,8 +3,8 @@ import { MDXRemote, Image, Giscus } from "@chia/components/client";
 import "highlight.js/styles/atom-one-dark-reasonable.css";
 import { Chip } from "@chia/components/server";
 import dayjs from "dayjs";
-import { serialize } from "@chia/utils/hydration.util";
 import { notFound } from "next/navigation";
+import type { Metadata } from "next";
 
 export const generateStaticParams = async () => {
   const posts = await getAllPosts();
@@ -17,15 +17,59 @@ export const generateStaticParams = async () => {
 // Global not found page
 // export const dynamicParams = false;
 
+export const generateMetadata = async ({
+  params,
+}: {
+  params: any;
+}): Promise<Metadata> => {
+  try {
+    const { frontMatter } = await getPost(params.slug);
+    return {
+      title: frontMatter?.title,
+      openGraph: {
+        type: "article",
+        locale: "zh_TW",
+        url: `https://chia1104.dev/posts/${params.slug}`,
+        siteName: "Chia",
+        title: frontMatter?.title,
+        description: frontMatter?.excerpt,
+        images: [
+          {
+            url: `https://chia1104.dev/api/og?title=${encodeURIComponent(
+              frontMatter?.title ?? ""
+            )}`,
+            width: 1200,
+            height: 630,
+          },
+        ],
+      },
+      twitter: {
+        card: "summary_large_image",
+        title: "Chia",
+        description: frontMatter?.excerpt,
+        creator: "@chia1104",
+        images: [
+          `https://chia1104.dev/api/og?title=${encodeURIComponent(
+            frontMatter?.title ?? ""
+          )}`,
+        ],
+      },
+    };
+  } catch (error) {
+    console.error(error);
+    return {};
+  }
+};
+
 const PostDetailPage = async ({ params }: { params?: any }) => {
   try {
     const { frontMatter, source } = await getPost(params.slug);
     return (
       <article className="main c-container mt-10 px-5">
-        <header className="pl-3 lg:w-[70%] w-full mb-7 self-center">
+        <header className="mb-7 w-full self-center pl-3 lg:w-[70%]">
           <h1 className="title pb-5">{frontMatter?.title}</h1>
           <h2 className="c-description">{frontMatter?.excerpt}</h2>
-          <span className="mt-5 flex items-center c-description gap-2">
+          <span className="c-description mt-5 flex items-center gap-2">
             <Image
               src="/memoji/contact-memoji.PNG"
               width={40}
@@ -38,10 +82,10 @@ const PostDetailPage = async ({ params }: { params?: any }) => {
           </span>
           <Chip data={frontMatter?.tags || []} />
         </header>
-        <div className="c-bg-secondary p-5 mt-5 rounded-xl lg:w-[70%] w-full self-center mx-auto">
-          <MDXRemote post={serialize(source)} />
+        <div className="c-bg-secondary mx-auto mt-5 w-full self-center rounded-xl p-5 lg:w-[70%]">
+          <MDXRemote post={source} />
         </div>
-        <div className="mt-20 lg:w-[70%] w-full self-center mx-auto">
+        <div className="mx-auto mt-20 w-full self-center lg:w-[70%]">
           <Giscus title={frontMatter?.title || ""} />
         </div>
       </article>
