@@ -9,7 +9,7 @@ const ContactButton: FC = () => {
 
   return (
     <button
-      className="mb-10 group hover:bg-secondary hover:dark:bg-primary relative inline-flex transition ease-in-out rounded"
+      className="group relative mb-10 inline-flex rounded transition ease-in-out hover:bg-secondary hover:dark:bg-primary"
       onClick={() => dispatch(activeActionIconSheet())}>
       <span className="c-button-secondary transform group-hover:-translate-x-1 group-hover:-translate-y-1">
         Contact me
