@@ -28,7 +28,7 @@ export class ErrorBoundary extends Component<Props, State> {
   public render() {
     if (this.state.hasError) {
       return (
-        <div className="w-screen h-screen c-container flex flex-col justify-center items-center">
+        <div className="c-container flex h-screen w-screen flex-col items-center justify-center">
           <h1 className="title text-warning">Oops, there is an error!</h1>
           <Image
             src="/error/error-memoji.png"
