@@ -2,7 +2,7 @@
 
 import { type FC, useState } from "react";
 import NextImage, { type ImageProps as NextImageProps } from "next/image";
-import cx from "classnames";
+import { cn } from "@chia//utils/cn.util";
 
 export interface ImageProps extends NextImageProps {
   blur?: boolean;
@@ -15,12 +15,12 @@ const Image: FC<ImageProps> = (props) => {
     <NextImage
       alt={alt}
       aria-label={alt}
-      className={cx(
+      className={cn(
         className,
         "duration-700 ease-in-out",
         isLoading
-          ? "grayscale blur-2xl scale-110"
-          : "grayscale-0 blur-0 scale-100"
+          ? "scale-110 blur-2xl grayscale"
+          : "scale-100 blur-0 grayscale-0"
       )}
       onLoadingComplete={() => setLoading(false)}
       {...rest}
