@@ -10,7 +10,7 @@ import React, {
   useRef,
 } from "react";
 import { ZodType } from "zod";
-import cx from "classnames";
+import { cn } from "@chia//utils/cn.util";
 
 interface Props
   extends DetailedHTMLProps<
@@ -89,8 +89,8 @@ const Input = forwardRef<InputRef, Props>((props, ref) => {
         type={type}
         onFocus={handleFocus}
         onBlur={handleBlur}
-        className={cx(
-          "border border-[#CBD2D7] w-full rounded-lg c-border-primary transition ease-in-out focus:outline-none c-bg-primary p-1",
+        className={cn(
+          "c-border-primary c-bg-primary w-full rounded-lg border border-[#CBD2D7] p-1 transition ease-in-out focus:outline-none",
           isError &&
             "border-danger hover:cursor-not-allowed dark:border-danger dark:hover:cursor-not-allowed",
           isFocus && !isError && "focus:border-info",
@@ -99,7 +99,7 @@ const Input = forwardRef<InputRef, Props>((props, ref) => {
         {...rest}
       />
       {isError && (
-        <p className={cx("text-danger", errorClassName)}>{error ?? ""}</p>
+        <p className={cn("text-danger", errorClassName)}>{error ?? ""}</p>
       )}
     </>
   );
