@@ -12,11 +12,10 @@ export { default as MDXRemote } from "./MDXRemote";
 export { default as Image } from "./Image";
 export { default as PostsList } from "./PostsList";
 export { default as Giscus } from "./Giscus";
-export { default as Head } from "./Head";
-export { default as DefaultSeo } from "./DefaultSeo";
 export { default as Analytics } from "./Analytics";
 export { default as RootProvider } from "./RootProvider";
 export { default as Input } from "./Input";
 export * from "./Input";
 export { default as Textarea } from "./Textarea";
 export * from "./Textarea";
+export * from "./Accordion";
