@@ -12,18 +12,18 @@ const Design: FC<Props> = ({ data }) => {
 
   return (
     <>
-      <header className="title sm:self-start c-text-bg-sec-half dark:c-text-bg-primary-half">
+      <header className="title c-text-bg-sec-half dark:c-text-bg-primary-half sm:self-start">
         Design
       </header>
-      <p className="c-description sm:self-start pb-7">Some of my design work</p>
+      <p className="c-description pb-7 sm:self-start">Some of my design work</p>
       <DMPosterList data={data} />
       <a
         href={POSTER_URL}
         target="_blank"
         rel="noopener noreferrer"
-        className="group hover:bg-secondary hover:dark:bg-primary relative inline-flex transition ease-in-out rounded mt-7 self-center"
+        className="group relative mt-7 inline-flex self-center rounded transition ease-in-out hover:bg-secondary hover:dark:bg-primary"
         aria-label={"Open Google Photos"}>
-        <span className="c-button-secondary transform group-hover:-translate-x-1 group-hover:-translate-y-1 text-base after:content-['_↗']">
+        <span className="c-button-secondary transform text-base after:content-['_↗'] group-hover:-translate-x-1 group-hover:-translate-y-1">
           Google Photos
         </span>
       </a>
