@@ -18,15 +18,15 @@ const GitHub: FC = asyncComponent(async () => {
 
     return (
       <>
-        <div className="w-full flex flex-col">
+        <div className="flex w-full flex-col">
           <ReposList repo={github} />
           <a
             href={`${GITHUB_URL}?tab=repositories`}
             target="_blank"
             rel="noopener noreferrer"
-            className="group hover:bg-secondary hover:dark:bg-primary relative inline-flex transition ease-in-out rounded mt-7 self-center"
+            className="group relative mt-7 inline-flex self-center rounded transition ease-in-out hover:bg-secondary hover:dark:bg-primary"
             aria-label="Open GitHub">
-            <span className="c-button-secondary transform group-hover:-translate-x-1 group-hover:-translate-y-1 text-base after:content-['_↗']">
+            <span className="c-button-secondary transform text-base after:content-['_↗'] group-hover:-translate-x-1 group-hover:-translate-y-1">
               GitHub
             </span>
           </a>
@@ -36,7 +36,7 @@ const GitHub: FC = asyncComponent(async () => {
   } catch (error) {
     console.error(error);
     return (
-      <div className="w-full flex flex-col">
+      <div className="flex w-full flex-col">
         <p className="c-description pb-5 indent-4">
           Sorry, I can't get my GitHub repos now. Please try again later.
         </p>
@@ -44,9 +44,9 @@ const GitHub: FC = asyncComponent(async () => {
           href={`${GITHUB_URL}?tab=repositories`}
           target="_blank"
           rel="noopener noreferrer"
-          className="group hover:bg-secondary hover:dark:bg-primary relative inline-flex transition ease-in-out rounded mt-7 self-center"
+          className="group relative mt-7 inline-flex self-center rounded transition ease-in-out hover:bg-secondary hover:dark:bg-primary"
           aria-label="Open GitHub">
-          <span className="c-button-secondary transform group-hover:-translate-x-1 group-hover:-translate-y-1 text-base after:content-['_↗']">
+          <span className="c-button-secondary transform text-base after:content-['_↗'] group-hover:-translate-x-1 group-hover:-translate-y-1">
             GitHub
           </span>
         </a>
