@@ -4,59 +4,59 @@ import { type FC } from "react";
 
 const ReposLoader: FC = () => {
   return (
-    <div className="w-full grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-10">
-      <div className="c-button-secondary w-full rounded h-[150px] flex flex-col">
-        <div className="mb-2 dark:bg-black bg-[#dddddd] animate-pulse h-5 rounded-full" />
-        <div className="line-clamp-2 dark:bg-black bg-[#dddddd] animate-pulse h-16 rounded-lg" />
+    <div className="grid w-full grid-cols-1 gap-10 md:grid-cols-2 xl:grid-cols-3">
+      <div className="c-button-secondary flex h-[150px] w-full flex-col rounded">
+        <div className="mb-2 h-5 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+        <div className="h-16 animate-pulse rounded-lg bg-[#dddddd] line-clamp-2 dark:bg-black" />
         <div className="mt-auto flex">
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
         </div>
       </div>
-      <div className="c-button-secondary w-full rounded h-[150px] flex flex-col">
-        <div className="mb-2 dark:bg-black bg-[#dddddd] animate-pulse h-5 rounded-full" />
-        <div className="line-clamp-2 dark:bg-black bg-[#dddddd] animate-pulse h-16 rounded-lg" />
+      <div className="c-button-secondary flex h-[150px] w-full flex-col rounded">
+        <div className="mb-2 h-5 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+        <div className="h-16 animate-pulse rounded-lg bg-[#dddddd] line-clamp-2 dark:bg-black" />
         <div className="mt-auto flex">
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
         </div>
       </div>
-      <div className="c-button-secondary w-full rounded h-[150px] flex flex-col">
-        <div className="mb-2 dark:bg-black bg-[#dddddd] animate-pulse h-5 rounded-full" />
-        <div className="line-clamp-2 dark:bg-black bg-[#dddddd] animate-pulse h-16 rounded-lg" />
+      <div className="c-button-secondary flex h-[150px] w-full flex-col rounded">
+        <div className="mb-2 h-5 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+        <div className="h-16 animate-pulse rounded-lg bg-[#dddddd] line-clamp-2 dark:bg-black" />
         <div className="mt-auto flex">
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
         </div>
       </div>
-      <div className="c-button-secondary w-full rounded h-[150px] flex flex-col">
-        <div className="mb-2 dark:bg-black bg-[#dddddd] animate-pulse h-5 rounded-full" />
-        <div className="line-clamp-2 dark:bg-black bg-[#dddddd] animate-pulse h-16 rounded-lg" />
+      <div className="c-button-secondary flex h-[150px] w-full flex-col rounded">
+        <div className="mb-2 h-5 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+        <div className="h-16 animate-pulse rounded-lg bg-[#dddddd] line-clamp-2 dark:bg-black" />
         <div className="mt-auto flex">
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
         </div>
       </div>
-      <div className="c-button-secondary w-full rounded h-[150px] flex flex-col">
-        <div className="mb-2 dark:bg-black bg-[#dddddd] animate-pulse h-5 rounded-full" />
-        <div className="line-clamp-2 dark:bg-black bg-[#dddddd] animate-pulse h-16 rounded-lg" />
+      <div className="c-button-secondary flex h-[150px] w-full flex-col rounded">
+        <div className="mb-2 h-5 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+        <div className="h-16 animate-pulse rounded-lg bg-[#dddddd] line-clamp-2 dark:bg-black" />
         <div className="mt-auto flex">
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
         </div>
       </div>
-      <div className="c-button-secondary w-full rounded h-[150px] flex flex-col">
-        <div className="mb-2 dark:bg-black bg-[#dddddd] animate-pulse h-5 rounded-full" />
-        <div className="line-clamp-2 dark:bg-black bg-[#dddddd] animate-pulse h-16 rounded-lg" />
+      <div className="c-button-secondary flex h-[150px] w-full flex-col rounded">
+        <div className="mb-2 h-5 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+        <div className="h-16 animate-pulse rounded-lg bg-[#dddddd] line-clamp-2 dark:bg-black" />
         <div className="mt-auto flex">
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="mr-5 dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
-          <span className="dark:bg-black bg-[#dddddd] animate-pulse h-3 w-10 rounded-full" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="mr-5 h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
+          <span className="h-3 w-10 animate-pulse rounded-full bg-[#dddddd] dark:bg-black" />
         </div>
       </div>
     </div>
