@@ -4,26 +4,26 @@ import type { FC } from "react";
 
 const VideoLoader: FC = () => {
   return (
-    <div className="w-full grid grid-cols-1 xl:grid-cols-2 gap-3">
-      <div className="w-full flex flex-col justify-center items-center">
-        <div className="dark:bg-dark bg-white animate-pulse h-5 mb-5 " />
-        <div className="dark:bg-dark bg-white animate-pulse w-full h-[270px] sm:h-[300px] sm:w-[500px] border-0 rounded-lg shadow-lg overflow-hidden mx-auto" />
+    <div className="grid w-full grid-cols-1 gap-3 xl:grid-cols-2">
+      <div className="flex w-full flex-col items-center justify-center">
+        <div className="mb-5 h-5 animate-pulse bg-white dark:bg-dark " />
+        <div className="mx-auto h-[270px] w-full animate-pulse overflow-hidden rounded-lg border-0 bg-white shadow-lg dark:bg-dark sm:h-[300px] sm:w-[500px]" />
       </div>
       <div className="my-5">
-        <div className="w-full flex flex-col c-border-primary border-b-2 p-3 h-[130px]">
-          <div className="dark:bg-dark bg-white animate-pulse h-5 mb-3" />
-          <div className="dark:bg-dark bg-white animate-pulse h-12" />
-          <div className="dark:bg-dark bg-white animate-pulse h-3 mt-auto w-40" />
+        <div className="c-border-primary flex h-[130px] w-full flex-col border-b-2 p-3">
+          <div className="mb-3 h-5 animate-pulse bg-white dark:bg-dark" />
+          <div className="h-12 animate-pulse bg-white dark:bg-dark" />
+          <div className="mt-auto h-3 w-40 animate-pulse bg-white dark:bg-dark" />
         </div>
-        <div className="w-full flex flex-col c-border-primary border-b-2 p-3 h-[130px]">
-          <div className="dark:bg-dark bg-white animate-pulse h-5 mb-3" />
-          <div className="dark:bg-dark bg-white animate-pulse h-12" />
-          <div className="dark:bg-dark bg-white animate-pulse h-3 mt-auto w-40" />
+        <div className="c-border-primary flex h-[130px] w-full flex-col border-b-2 p-3">
+          <div className="mb-3 h-5 animate-pulse bg-white dark:bg-dark" />
+          <div className="h-12 animate-pulse bg-white dark:bg-dark" />
+          <div className="mt-auto h-3 w-40 animate-pulse bg-white dark:bg-dark" />
         </div>
-        <div className="w-full flex flex-col c-border-primary border-b-2 p-3 h-[130px]">
-          <div className="dark:bg-dark bg-white animate-pulse h-5 mb-3" />
-          <div className="dark:bg-dark bg-white animate-pulse h-12" />
-          <div className="dark:bg-dark bg-white animate-pulse h-3 mt-auto w-40" />
+        <div className="c-border-primary flex h-[130px] w-full flex-col border-b-2 p-3">
+          <div className="mb-3 h-5 animate-pulse bg-white dark:bg-dark" />
+          <div className="h-12 animate-pulse bg-white dark:bg-dark" />
+          <div className="mt-auto h-3 w-40 animate-pulse bg-white dark:bg-dark" />
         </div>
       </div>
     </div>
