@@ -9,7 +9,7 @@ const Youtube: FC = asyncComponent(async () => {
   try {
     const { status, data } = await getAllVideos(4);
     return (
-      <div className="w-full flex flex-col">
+      <div className="flex w-full flex-col">
         {status === 200 ? (
           <VideoList item={data.items} />
         ) : (
@@ -21,9 +21,9 @@ const Youtube: FC = asyncComponent(async () => {
           href={YOUTUBE_URL}
           target="_blank"
           rel="noopener noreferrer"
-          className="group hover:bg-secondary hover:dark:bg-primary relative inline-flex transition ease-in-out rounded mt-7 self-center"
+          className="group relative mt-7 inline-flex self-center rounded transition ease-in-out hover:bg-secondary hover:dark:bg-primary"
           aria-label={"Open Youtube"}>
-          <span className="c-button-secondary transform group-hover:-translate-x-1 group-hover:-translate-y-1 text-base after:content-['_↗']">
+          <span className="c-button-secondary transform text-base after:content-['_↗'] group-hover:-translate-x-1 group-hover:-translate-y-1">
             Youtube
           </span>
         </a>
@@ -32,7 +32,7 @@ const Youtube: FC = asyncComponent(async () => {
   } catch (error) {
     console.error(error);
     return (
-      <div className="w-full flex flex-col">
+      <div className="flex w-full flex-col">
         <p className="c-description pb-5 indent-4">
           Sorry, I can't get my Youtube videos now. Please try again later.
         </p>
@@ -40,9 +40,9 @@ const Youtube: FC = asyncComponent(async () => {
           href={YOUTUBE_URL}
           target="_blank"
           rel="noopener noreferrer"
-          className="group hover:bg-secondary hover:dark:bg-primary relative inline-flex transition ease-in-out rounded mt-7 self-center"
+          className="group relative mt-7 inline-flex self-center rounded transition ease-in-out hover:bg-secondary hover:dark:bg-primary"
           aria-label={"Open Youtube"}>
-          <span className="c-button-secondary transform group-hover:-translate-x-1 group-hover:-translate-y-1 text-base after:content-['_↗']">
+          <span className="c-button-secondary transform text-base after:content-['_↗'] group-hover:-translate-x-1 group-hover:-translate-y-1">
             Youtube
           </span>
         </a>
